refactor: 정적 팩토리 메소드 파라미터 변경 및 사용하지 않는 메서드 제거
