+++ src/main/java/chess/domain/board/Direction.java
@@ -1,6 +1,5 @@
-package chess.domain.position;
+package chess.domain.board;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
@@ -29,7 +28,6 @@ public enum Direction {
     public static final List<Direction> KNIGHT_DIRECTION = Arrays.asList(NORTH_NORTH_EAST, NORTH_NORTH_WEST, SOUTH_SOUTH_EAST, SOUTH_SOUTH_WEST, EAST_EAST_NORTH, EAST_EAST_SOUTH, WEST_WEST_NORTH, WEST_WEST_SOUTH);
     public static final List<Direction> WHITE_PAWN_DIRECTION = Arrays.asList(NORTH, NORTH_EAST, NORTH_WEST);
     public static final List<Direction> BLACK_PAWN_DIRECTION = Arrays.asList(SOUTH, SOUTH_EAST, SOUTH_WEST);
-    private static final int NEXT = 1;
 
     private final int x;
     private final int y;
@@ -39,14 +37,11 @@ public enum Direction {
         this.y = y;
     }
 
-    public static Direction of(final Position from, final Position to) {
-        final var xDifference = to.getFileOrder() - from.getFileOrder();
-        final var yDifference = to.getRankNumber() - from.getRankNumber();
-
+    public static Direction of(final int xDifference, final int yDifference) {
         if (isLine(xDifference, yDifference)) {
-            return of(calculateLineDifference(xDifference), calculateLineDifference(yDifference));
+            return findDirection(calculateLineDifference(xDifference), calculateLineDifference(yDifference));
         }
-        return of(xDifference, yDifference);
+        return findDirection(xDifference, yDifference);
     }
 
     private static boolean isLine(final int xDifference, final int yDifference) {
@@ -68,7 +63,7 @@ public enum Direction {
         return difference / Math.abs(difference);
     }
 
-    private static Direction of(final int xDifference, final int yDifference) {
+    private static Direction findDirection(final int xDifference, final int yDifference) {
         return Arrays.stream(values())
                 .filter(direction -> direction.x == xDifference && direction.y == yDifference)
                 .findFirst()
@@ -83,16 +78,11 @@ public enum Direction {
         return this == Direction.NORTH_EAST || this == Direction.SOUTH_EAST || this == Direction.SOUTH_WEST || this == Direction.NORTH_WEST;
     }
 
-    public List<Position> getPositions(final Position start, final Position end) {
-        final var rankDifference = Math.abs(start.getRankNumber() - end.getRankNumber());
-        final var fileDifference = Math.abs(start.getFileOrder() - end.getFileOrder());
-        final var endCount = Math.max(rankDifference, fileDifference);
-        var file = start.getFileOrder() + x;
-        var rank = start.getRankNumber() + y;
-        final List<Position> positions = new ArrayList<>();
-        for (int i = NEXT; i < endCount; i++, file += x, rank += y) {
-            positions.add(Position.of(File.findByOrder(file), Rank.from(rank)));
-        }
-        return positions;
+    public int getX() {
+        return x;
+    }
+
+    public int getY() {
+        return y;
     }
 }
